Update readme
Here, we consider two different simulated data sets. The first one is based on a constant value for baseline hazard, so the inverse of the cumulative hazard function is a common approach for this purpose. The R code can be found in "simulated_JM1.R" and the simulated data can be found in "long.data_1.RData" and "surv.data_1.RData". The second one is based on the permutational algorithm using the *PermAlgo* package. The R code for this method can be found in "simulated_JM2.R" and the simulated data can be found in "long.data_2.RData" and "surv.data_2..RData". 